// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/custom-field-history.model.ts
export class CustomFieldHistoryModel {
    customFieldId: string;
    fieldName: string;
    formJson: string;
    formDataJson: string;
    referenceId: string;
    referenceTypeId: string;
    oldValue: string;
    newValue: string;
    description: string;
    fullName: string;
    profileImage: string;
    createdDateTime: Date;
    customFieldHistoryId: string;
}